Refactor getVarInits
